+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added SDSR loop and Evoloop implementations
+- Added `memoize` option to `evolve` and `evolve2d` functions
 
 ## [2.1.0] - 2021-11-16